feat: запуск по cron
