+++ package.json
@@ -5,6 +5,7 @@
   "main": "dist/app.js",
   "scripts": {
     "start": "tsc && node dist/app.js",
+    "build": "tsc",
     "lint": "eslint . --ext .ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -24,6 +25,7 @@
   },
   "dependencies": {
     "axios": "^1.4.0",
+    "cron": "^2.4.3",
     "csv-writer": "^1.6.0",
     "dotenv": "^16.3.1",
     "json-2-csv": "^4.1.0",
